feat: Tech community landing page updated.
