+++ frontend/src/Pages/SingleTechCommunity.jsx
@@ -50,6 +50,7 @@ import getTimeAgo from "../components/DateCovertion";
 import useGetAllMembersByDomain from "../hooks/SingleTechDomain/useGetAllMembersByDomain";
 // import coordinatorsCard from "../components/authors/CoordinatorsCard";
 import CoordinatorsCard from "../components/authors/CoordinatorsCard";
+import { deriveGradient } from "../utils/bannerTheme";
 
 // ── S3 image base ─────────────────────────────────────────────────────────────
 const S3 = "https://open-access-blog-image.s3.us-east-1.amazonaws.com/";
@@ -121,53 +122,6 @@ const timeAgo = (dateStr) => {
 // ── Community banner ──────────────────────────────────────────────────────────
 const CommunityBanner = ({ community, style, communityId, canEdit }) => {
   
-
-  const hexToHsl = (hex) => {
-  const r = parseInt(hex.slice(1, 3), 16) / 255;
-  const g = parseInt(hex.slice(3, 5), 16) / 255;
-  const b = parseInt(hex.slice(5, 7), 16) / 255;
-  const max = Math.max(r, g, b), min = Math.min(r, g, b);
-  let h, s, l = (max + min) / 2;
-  if (max === min) { h = s = 0; }
-  else {
-    const d = max - min;
-    s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
-    switch (max) {
-      case r: h = ((g - b) / d + (g < b ? 6 : 0)) / 6; break;
-      case g: h = ((b - r) / d + 2) / 6; break;
-      case b: h = ((r - g) / d + 4) / 6; break;
-    }
-  }
-  return [Math.round(h * 360), Math.round(s * 100), Math.round(l * 100)];
-};
-
-const hslToHex = (h, s, l) => {
-  h /= 360; s /= 100; l /= 100;
-  const hue2rgb = (p, q, t) => {
-    if (t < 0) t += 1; if (t > 1) t -= 1;
-    if (t < 1 / 6) return p + (q - p) * 6 * t;
-    if (t < 1 / 2) return q;
-    if (t < 2 / 3) return p + (q - p) * (2 / 3 - t) * 6;
-    return p;
-  };
-  const q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-  const p = 2 * l - q;
-  const r = Math.round(hue2rgb(p, q, h + 1 / 3) * 255);
-  const g = Math.round(hue2rgb(p, q, h) * 255);
-  const b = Math.round(hue2rgb(p, q, h - 1 / 3) * 255);
-  return `#${[r, g, b].map((x) => x.toString(16).padStart(2, "0")).join("")}`;
-};
-
-  const deriveGradient = (baseHex) => {
-    if (!baseHex || !/^#[0-9A-Fa-f]{6}$/.test(baseHex)) {
-      return { from: "#0d9488", to: "#0f766e" };
-    }
-    const [h, s, l] = hexToHsl(baseHex);
-    return {
-      from: baseHex,
-      to: hslToHex(h, Math.min(s + 5, 100), Math.max(l - 15, 5)),
-    };
-  };
 
   const gradient = useMemo(() => {
     const theme = community?.colorTheme;
